+++ jackson/src/test/java/org/fuin/objects4j/jackson/DayOpeningHoursJacksonSerializerTest.java
@@ -26,7 +26,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 /**
- * Test for the {@link DayOpeningHoursJacksonSerializer} class.
+ * Test for {@link DayOpeningHours} serialization.
  */
 public class DayOpeningHoursJacksonSerializerTest {