+++ jackson/src/test/java/org/fuin/objects4j/jackson/EmailAddressJacksonDeserializerTest.java
@@ -26,7 +26,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 /**
- * Test for the {@link EmailAddressJacksonDeserializer} class.
+ * Test for the {@link EmailAddress} deserialization.
  */
 public class EmailAddressJacksonDeserializerTest {