+++ jackson/src/test/java/org/fuin/objects4j/jackson/HourJacksonDeserializerTest.java
@@ -26,7 +26,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 /**
- * Test for the {@link HourJacksonDeserializer} class.
+ * Test for {@link Hour} deserialization.
  */
 public class HourJacksonDeserializerTest {